[Refactor] : Join과 Profile 분리
