+++ src/main/java/com/potato/balbambalbam/user/profile/dto/EditResponseDto.java
@@ -1,4 +1,4 @@
-package com.potato.balbambalbam.user.join.dto;
+package com.potato.balbambalbam.user.profile.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
